Write createPlayer() and styling for index element
createPlayer function inside a module and styling for the names of the
players. This way this is done after the user push the submit button.
Also added the images for picard, janeway and the borg queen.
Trying to make everything work inside functions or inside modules.
Avoiding free code all together.