Modify the player factory function
This commit adds a little bit of functionality to the object so it gets
easier to send data to the dom, and to the 'const game'. That way I may
use the new methods to check for victory and announce the winner, also
adding points to the score of the game.