Finish Human-only gameplay experience
All the functionalities for the human only experience are working.
script.js has now the victory, checkforvictory and tie functions.
The only external part of code is the currentPlayer variable. Everything
else is inside a function or a module.

Also this commit finish the personalLog entries and the documentation part.

Next step is pushing it to the repository and publishing in TOP
solutions.